auth added

+++ frontend/src/components/WorkoutDetails.jsx
@@ -2,16 +2,31 @@ import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
 import { MdDelete } from "react-icons/md";
 import  formatDistanceToNow  from 'date-fns/formatDistanceToNow';
 import toast from "react-hot-toast";
-
+import { useAuthContext } from "../hooks/useAuthContext";
 const WorkoutDetails = ({ workout }) => {
+  const {user} = useAuthContext();
   const {dispatch} = useWorkoutsContext();
 const handleClick = async ()=>{
+  if(!user){
+    return ;
+  }
    const response = await fetch("/api/workouts/"+workout._id,{
-    method:'DELETE'
+    method:'DELETE',
+    headers:{
+      'Authorization':`Bearer ${user.token}`
+
+    }
    });
    const json = await response.json();
     if(response.ok){
-      toast.success("Workout Removed");
+      toast.success("Workout Removed", {
+        icon: '👏',
+        style: {
+          borderRadius: '10px',
+          background: '#333',
+          color: '#fff',
+        },
+      });
       dispatch({type:'DELETE_WORKOUT',payload:json})
     }
 }
